requirements.txt
python-telegram-bot==21.10
python-dotenv==0.19.0
fuzzywuzzy==0.18.0
Pillow==10.2.0
pytz==2024.1
cairosvg==2.7.1
psutil==5.9.6
requests==2.31.0
aiohttp==3.9.1
ujson==5.8.0
firebase-admin==6.4.0
google-analytics-admin==0.18.0
google-analytics-data==0.18.0
python-Levenshtein==0.23.0
tabulate==0.9.0